feat: remove fun subpage, refactor code
Removed fun subpage, and all links to it.
Cleanup router.
Added helmet for security headers.
Updated dependencies.
