+++ package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "async": "2.6.2",
     "basic-auth": "^2.0.1",
-    "bitcoin-core": "2.0.0",
+    "bitcoin-core": "^1.2.0",
     "bitcoinjs-lib": "5.0.3",
     "body-parser": "~1.19.0",
     "compression": "^1.7.4",
@@ -37,6 +37,7 @@
     "electrum-client": "git+https://git@github.com/chaintools/node-electrum-client.git#43a999036f9c5",
     "express": "^4.17.1",
     "express-session": "1.16.1",
+    "helmet": "^4.6.0",
     "jstransformer-markdown-it": "^2.1.0",
     "lru-cache": "5.1.1",
     "md5": "^2.3.0",
@@ -46,7 +47,7 @@
     "morgan": "^1.9.1",
     "pug": "^3.0.2",
     "qrcode": "^1.4.4",
-    "redis": "2.8.0",
+    "redis": "^3.1.2",
     "request": "2.88.0",
     "serve-favicon": "^2.5.0",
     "simple-git": "1.112.0"
