+++ views/address.pug
@@ -29,41 +29,6 @@ block content
 					p
 						span This looks like a miner payout address for  
 						span(class="font-weight-bold") #{payoutAddressForMiner.name}
-			else if (global.specialAddresses[address] && global.specialAddresses[address].type == "donation")
-				div(class="alert alert-primary shadow-sm", style="padding-bottom: 0;")
-					div(class="float-left", style="width: 50px; height: 50px; font-size: 18px;")
-						i(class="fas fa-certificate fa-2x", style="margin-top: 10px;")
-
-					h4(class="alert-heading h6 font-weight-bold") #{coinConfig.name} Fun
-
-					p
-						span This is the #{coinConfig.ticker} donation address to support development of this tool. All support is appreciated!
-
-			else if (global.specialAddresses[address] && global.specialAddresses[address].type == "fun")
-				div(class="alert alert-primary shadow-sm", style="padding-bottom: 0;")
-					div(class="float-left", style="width: 50px; height: 50px; font-size: 18px;")
-						i(class="fas fa-certificate fa-2x", style="margin-top: 10px;")
-
-					h4(class="alert-heading h6 font-weight-bold") #{coinConfig.name} Fun
-
-					// special address info
-					- var saInfo = global.specialAddresses[address].addressInfo;
-					if (saInfo.alertBodyHtml)
-						p
-							span !{saInfo.alertBodyHtml}
-
-							if (saInfo.referenceUrl && saInfo.referenceUrl.trim().length > 0 && saInfo.alertBodyHtml.indexOf(saInfo.referenceUrl) == -1)
-								span  
-									a(href=saInfo.referenceUrl) Read more
-
-					else
-						p
-							span #{saInfo.summary}
-
-							if (saInfo.referenceUrl && saInfo.referenceUrl.trim().length > 0)
-								span  
-									a(href=saInfo.referenceUrl) Read more
-
 
 			ul(class='nav nav-tabs tabs')
 				li(class="nav-item tabs-item")
@@ -295,11 +260,6 @@ block content
 											div(class="col-md-6")
 												if (tx && tx.txid)
 													a(href=("/tx/" + tx.txid), class="blue-link") #{tx.txid}
-
-													if (global.specialTransactions && global.specialTransactions[tx.txid])
-														span
-														a(data-toggle="tooltip", title=(coinConfig.name + " Fun! See transaction for details"))
-															i(class="fas fa-certificate text-primary")
 
 											div(class="col-md-6")
 												div(class="text-md-right")
