+++ views/includes/line-graph.pug
@@ -1,4 +1,4 @@
-script(src="/js/chart.bundle.min.js", integrity="sha384-e4YKd0O/y4TmH7qskMQzKnOrqN83RJ7TmJ4RsBLHodJ6jHOE30I7J1uZfLdvybhc")
+script(src="/js/chart.bundle.min.js")
 
 canvas(id=graphData.id, class="mb-4")
