+++ views/includes/transaction-io-details.pug
@@ -45,16 +45,6 @@ div(class="row")
 										if (vout && vout.scriptPubKey && vout.scriptPubKey.addresses)
 											div(class="word-wrap")
 												a(href=("/address/" + vout.scriptPubKey.addresses[0]), class=("blue-link")) #{vout.scriptPubKey.addresses[0]}
-												if (global.specialAddresses[vout.scriptPubKey.addresses[0]])
-													- var specialAddressInfo = global.specialAddresses[vout.scriptPubKey.addresses[0]];
-													if (specialAddressInfo.type == "minerPayout")
-														span  
-														a(data-toggle="tooltip", title=("Miner payout address: " + specialAddressInfo.minerInfo.name))
-															i(class="fas fa-certificate text-primary")
-													else if (specialAddressInfo.type == "donation")
-														span  
-														a(data-toggle="tooltip", title=("Development donation address. All support is appreciated!"))
-															i(class="fas fa-certificate text-primary")
 
 											span(class="small") via 
 												a(class="blue-link", href=("/tx/" + txInput.txid + "#output-" + txVin.vout)) #{txInput.txid.substring(0, 20)}...[#{txVin.vout}]
@@ -139,18 +129,6 @@ div(class="row")
 
 										a(id=("output-" + voutIndex), href=("/address/" + vout.scriptPubKey.addresses[0]), class="blue-link")
 											span(class="word-wrap ") #{vout.scriptPubKey.addresses[0]}
-
-										if (global.specialAddresses[vout.scriptPubKey.addresses[0]])
-											- var specialAddressInfo = global.specialAddresses[vout.scriptPubKey.addresses[0]];
-											if (specialAddressInfo.type == "minerPayout")
-												span  
-												a(data-toggle="tooltip", title=("Miner payout address: " + specialAddressInfo.minerInfo.name))
-													i(class="fas fa-certificate text-primary")
-											else if (specialAddressInfo.type == "donation")
-												span  
-												a(data-toggle="tooltip", title=("Development donation address. All support is appreciated!"))
-													i(class="fas fa-certificate text-primary")
-													
 
 									else if (vout.scriptPubKey.hex && vout.scriptPubKey.hex.startsWith('6a24aa21a9ed'))
 										span()
