+++ views/layout.pug
@@ -6,14 +6,13 @@ html(lang="en")
 		meta(name="viewport", content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=0, shrink-to-fit=no")
 		
 		link(rel="stylesheet", href="/css/fonts.css")
-		link(rel="stylesheet", href="/css/highlight.min.css", integrity="sha384-zhIsEafzyQWHSoMCQ4BfT8ZlRXQyIFwAHAJn32PNdsb8n6tVysGZSLpEEIvCskw4")
+		link(rel="stylesheet", href="/css/highlight.min.css")
 
 		if (session.uiTheme && session.uiTheme == "dark")
-			link(rel="stylesheet", href="/css/bootstrap-dark.css", integrity="")
-			link(rel="stylesheet", href="/css/dark-touchups.css", integrity="")
-				
+			link(rel="stylesheet", href="/css/bootstrap-dark.css")
+			link(rel="stylesheet", href="/css/dark-touchups.css")
 		else
-			link(rel="stylesheet", href="/css/bootstrap.min.css", integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T")
+			link(rel="stylesheet", href="/css/bootstrap.min.css")
 
 		
 		link(rel='stylesheet', href='/css/styling.css')
@@ -160,9 +159,9 @@ html(lang="en")
 
 				
 		script(src="/js/jquery.min.js", type="text/javascript")
-		script(src="/js/popper.min.js", integrity="sha384-wHAiFfRlMFy6i5SRaxvfOCifBUQy1xHdJ/yoi7FRNXMRBu5WHdZYu1hA6ZOblgut")
-		script(src="/js/bootstrap.min.js", integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM")
-		script(defer, src="/js/fontawesome.min.js", integrity="sha384-eVEQC9zshBn0rFj4+TU78eNA19HMNigMviK/PU/FFjLXqa/GKPgX58rvt5Z8PLs7")
+		script(src="/js/popper.min.js")
+		script(src="/js/bootstrap.min.js")
+		script(defer, src="/js/fontawesome.min.js")
 		script(src="/js/highlight.min.js", charset="UTF-8")
 		
 		script.
