+++ views/mining-pools.pug
@@ -1,7 +1,7 @@
 extends layout
 
 block headContent
-	script(src="/js/chart.bundle.min.js", integrity="sha384-e4YKd0O/y4TmH7qskMQzKnOrqN83RJ7TmJ4RsBLHodJ6jHOE30I7J1uZfLdvybhc")
+	script(src="/js/chart.bundle.min.js")
 	title Hashrate distribution
 
 block content
